[TEST] new tests have been added. Tests have been updated and documentation has been improved.
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/branches/msnovogen@11678 6adb6e08-d915-0410-941f-83917bcadc18

+++ source/ANALYSIS/DENOVO/MSNOVOGEN/Mater.C
@@ -44,7 +44,7 @@ namespace OpenMS
 	Mater::Mater(double precursorMass, double precursorMassTolerance, std::vector<const Residue*> aaList) :
 	aaList_(aaList), precursorMass_(precursorMass), precursorMassTolerance_(precursorMassTolerance)
 	{
-	  seed(time(0));
+	  seed((unsigned int)time(0));
 	}
 
 	Mater::~Mater()
